+++ projs/shadow/shadow-engine/core/src/core/Time.cpp
@@ -5,8 +5,8 @@ int Time::NOW = 0;//SDL_GetPerformanceCounter();
 int Time::LAST = 0;
 double Time::deltaTime_ms = 0;
 double Time::deltaTime = 0;
-size_t Time::startTime = 0;
-size_t Time::timeSinceStart = 0;
+double Time::startTime = 0;
+double Time::timeSinceStart = 0;
 
 void Time::UpdateTime()
 {
@@ -15,7 +15,7 @@ void Time::UpdateTime()
     auto now_ms = time_point_cast<milliseconds>(now);
 
     auto value = now_ms.time_since_epoch();
-    size_t duration = value.count();
+    double duration = value.count();
 
     if (startTime == 0)
         startTime = duration;
